Move pymc3.model.pandas_to_array to pymc3.aesaraf
