+++ pymc3/tests/test_aesaraf.py
@@ -17,14 +17,25 @@ from itertools import product
 import aesara
 import aesara.tensor as aet
 import numpy as np
+import numpy.ma as ma
+import numpy.testing as npt
+import pandas as pd
 import pytest
+import scipy.sparse as sps
 
+from aesara.graph.basic import Variable
 from aesara.tensor.subtensor import AdvancedIncSubtensor, AdvancedIncSubtensor1
 from aesara.tensor.type import TensorType
+from aesara.tensor.var import TensorVariable
 
 import pymc3 as pm
 
-from pymc3.aesaraf import _conversion_map, extract_obs_data, take_along_axis
+from pymc3.aesaraf import (
+    _conversion_map,
+    extract_obs_data,
+    pandas_to_array,
+    take_along_axis,
+)
 from pymc3.vartypes import int_types
 
 FLOATX = str(aesara.config.floatX)
@@ -302,3 +313,83 @@ def test_extract_obs_data():
 
     assert isinstance(res, np.ndarray)
     assert np.ma.allequal(res, data_m)
+
+
+@pytest.mark.parametrize("input_dtype", ["int32", "int64", "float32", "float64"])
+def test_pandas_to_array(input_dtype):
+    """
+    Ensure that pandas_to_array returns the dense array, masked array,
+    graph variable, TensorVariable, or sparse matrix as appropriate.
+    """
+    # Create the various inputs to the function
+    sparse_input = sps.csr_matrix(np.eye(3)).astype(input_dtype)
+    dense_input = np.arange(9).reshape((3, 3)).astype(input_dtype)
+
+    input_name = "input_variable"
+    aesara_graph_input = aet.as_tensor(dense_input, name=input_name)
+    pandas_input = pd.DataFrame(dense_input)
+
+    # All the even numbers are replaced with NaN
+    missing_numpy_input = np.array([[np.nan, 1, np.nan], [3, np.nan, 5], [np.nan, 7, np.nan]])
+    missing_pandas_input = pd.DataFrame(missing_numpy_input)
+    masked_array_input = ma.array(dense_input, mask=(np.mod(dense_input, 2) == 0))
+
+    # Create a generator object. Apparently the generator object needs to
+    # yield numpy arrays.
+    square_generator = (np.array([i ** 2], dtype=int) for i in range(100))
+
+    # Alias the function to be tested
+    func = pandas_to_array
+
+    #####
+    # Perform the various tests
+    #####
+    # Check function behavior with dense arrays and pandas dataframes
+    # without missing values
+    for input_value in [dense_input, pandas_input]:
+        func_output = func(input_value)
+        assert isinstance(func_output, np.ndarray)
+        assert func_output.shape == input_value.shape
+        npt.assert_allclose(func_output, dense_input)
+
+    # Check function behavior with sparse matrix inputs
+    sparse_output = func(sparse_input)
+    assert sps.issparse(sparse_output)
+    assert sparse_output.shape == sparse_input.shape
+    npt.assert_allclose(sparse_output.toarray(), sparse_input.toarray())
+
+    # Check function behavior when using masked array inputs and pandas
+    # objects with missing data
+    for input_value in [missing_numpy_input, masked_array_input, missing_pandas_input]:
+        func_output = func(input_value)
+        assert isinstance(func_output, ma.core.MaskedArray)
+        assert func_output.shape == input_value.shape
+        npt.assert_allclose(func_output, masked_array_input)
+
+    # Check function behavior with Aesara graph variable
+    aesara_output = func(aesara_graph_input)
+    assert isinstance(aesara_output, Variable)
+    npt.assert_allclose(aesara_output.eval(), aesara_graph_input.eval())
+    intX = pm.aesaraf._conversion_map[aesara.config.floatX]
+    if dense_input.dtype == intX or dense_input.dtype == aesara.config.floatX:
+        assert aesara_output.owner is None  # func should not have added new nodes
+        assert aesara_output.name == input_name
+    else:
+        assert aesara_output.owner is not None  # func should have casted
+        assert aesara_output.owner.inputs[0].name == input_name
+
+    if "float" in input_dtype:
+        assert aesara_output.dtype == aesara.config.floatX
+    else:
+        assert aesara_output.dtype == intX
+
+    # Check function behavior with generator data
+    generator_output = func(square_generator)
+
+    # Output is wrapped with `pm.floatX`, and this unwraps
+    wrapped = generator_output.owner.inputs[0]
+    # Make sure the returned object has .set_gen and .set_default methods
+    assert hasattr(wrapped, "set_gen")
+    assert hasattr(wrapped, "set_default")
+    # Make sure the returned object is a Aesara TensorVariable
+    assert isinstance(wrapped, TensorVariable)
